// src/app/entry/interface/output-jobdata.interface.ts
// Generated by https://quicktype.io

export interface OutputJobData {
  DeletedByOPID?:      string;
  DeletedOnSessionID?: string;
  JobDataInfo?:        JobDataInfo;
  Name:               string;
  PageDatas:          PageData[];
  DeletedPageDatas?:   any[];
}

export interface JobDataInfo {
  BatchName:            string;
  Customer:             string;
  DateTimeIn:           string;
  DateTimeOut:          string;
  DCN:                  string;
  DocumentClosedByFlag: boolean;
  DocumentFlagMessage:  null;
  DocumentsCount:       number;
  ElapsedTime:          number;
  IpAddress:            string;
  Job:                  string;
  KeyGross:             number;
  KeyNet:               number;
  LinesCount:           number;
  MachineName:          string;
  MachineUser:          string;
  OPID:                 string;
  PauseTime:            number;
  Process:              string;
  ReferenceID:          number;
  RoutePath:            string;
  SessionId:            string;
  Status:               string;
  Version:              string;
}

export interface PageData {
  DeletedByOPID?:      string;
  DeletedOnSessionID?: string;
  GroupDatas:         GroupData[];
  DeletedGroupDatas?:  any[];
  IsNewElement:       boolean;
  Name:               string;
}

export interface GroupData {
  DeletedByOPID?:      string;
  DeletedOnSessionID?: string;
  Fields?:             FieldData[];
  IsNewElement:       boolean;
  Name:               string;
}

export interface FieldData {
  CurrentPosition?: string | null;
  LocationInfo?:    { [key: string]: number } | null;
  FlagMessage:     null;
  InputData?:       InputData;
  IsNewElement:    boolean;
  KeyGross:        number;
  KeyNet:          number;
  Name:            string;
  OPID:            string;
  PageIndex:       number;
  SessionID:       string;
  Status:          string;
  Type:            string;
  Value:           string;
  Zoom?:            string;
}

export interface InputData {
  CurrentPosition: null;
  Name:            null;
  PageIndex:       number;
  Status:          null;
  Type:            null;
  Values:          any[];
  ZoneHighlight:   null;
  Zoom:            null;
  LocationInfo:    null;
}