// src/app/entry/interface/input-jobdata.interface.ts
// Generated by https://quicktype.io

export interface InputJobData {
  ImageListIndex: any[];
  JobDataInfo:    InputJobDataInfo;
  Name:           string;
  PageDatas:      InputPageData[];
}

export interface InputJobDataInfo {
  BatchName:            null | string;
  Customer:             string;
  DateTimeIn:           string;
  DateTimeOut:          string;
  DCN:                  string;
  DocumentClosedByFlag: boolean;
  DocumentFlagMessage:  string;
  DocumentsCount:       number;
  ElapsedTime:          number;
  IpAddress:            string;
  Job:                  string;
  KeyGross:             number;
  KeyNet:               number;
  LinesCount:           number;
  MachineName:          string;
  MachineUser:          string;
  OPID:                 string;
  PauseTime:            number;
  Process:              string;
  ReferenceID:          number;
  RoutePath:            string;
  SessionId:            null;
  Status:               string;
  Version:              string;
}

export interface InputPageData {
  GroupDatas: InputGroupData[];
  Name:       string;
}

export interface InputGroupData {
  Fields: InputFieldData[];
  Name:   string;
}

export interface InputFieldData {
  CurrentPosition: CurrentPosition | null;
  Name:            string;
  PageIndex:       number;
  Status:          string;
  Type:            string;
  Values:          InputValue[];
  ZoneHighlight:   null;
  Zoom:            CurrentPosition | null;
  LocationInfo:    null;
}

export interface CurrentPosition {
  X: number;
  Y: number;
}

export interface InputValue {
  Accuracy:    number;
  Content:     string;
  FlagMessage: null;
  OPID:        string;
  Process:     string;
  SessionID:   string;
}